try to fix issues in provisioning library tests
